пагінація працює всюди крім країн
