+++ src/js/services/services.js
@@ -1,18 +1,20 @@
 const API_KEY = 'MMQ2M3AOTcNvFmVoIxNGUGotXqF5t9MP';
-const BASE_URL = 'https://newsuperserver.herokuapp.com/https://app.ticketmaster.com';
+const BASE_URL = 'https://app.ticketmaster.com';
 
 export default class EventsApiService {
     constructor() {
         this.searchQuery = '';
         this.page = 1;
+        this.size = 20;
+
     }
 
     async fetchEvent() {
         let url = '';
         if (this.searchQuery !== '') {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&&keyword=${this.searchQuery}apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&keyword=${this.searchQuery}&apikey=${API_KEY}`;
         } else {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&apikey=${API_KEY}`;
         }
 
         const rawResult = await fetch(url);
@@ -25,12 +27,13 @@ export default class EventsApiService {
         // зразу повертаємо масив подій
         return result._embedded.events;
     }
+
     async fetchPages() {
         let url = '';
         if (this.searchQuery !== '') {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&&keyword=${this.searchQuery}apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&keyword=${this.searchQuery}&apikey=${API_KEY}`;
         } else {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&apikey=${API_KEY}`;
         }
         const rawResult = await fetch(url);
         if (!rawResult.ok) {
@@ -46,9 +49,9 @@ export default class EventsApiService {
     async fetchNextEvent() {
         let url = '';
         if (this.searchQuery === '') {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&page=${this.page}&apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&page=${this.page}&apikey=${API_KEY}`;
         } else {
-            url = `${BASE_URL}/discovery/v2/events.json?size=20&keyword=${this.searchQuery}&page=${this.page}&apikey=${API_KEY}`;
+            url = `${BASE_URL}/discovery/v2/events.json?size=${this.size}&keyword=${this.searchQuery}&page=${this.page}&apikey=${API_KEY}`;
         }
         console.log(url);
         const rawResult = await fetch(url);
@@ -63,8 +66,14 @@ export default class EventsApiService {
         return result._embedded.events;
 
     }
+    changeCountry(newCountry) {
+        this.countryCode = newCountry;
+    }
     changePage(newPage) {
         this.page = newPage;
+    }
+    changeSize(newSize) {
+        this.size = newSize;
     }
