+++ src/js/services/start-page.js
@@ -3,7 +3,7 @@ const cardEvent = document.querySelector('.gallery');
 const eventsApiService = new EventsApiService();
 import eventsListTpl from '../../templates/card-list.hbs';
 import createsDownloadList from './nev-render-list.js';
-import getPage from './get-page.js';
+//import getPage from './get-page.js';
 
 function onSearch() {
     try {
@@ -24,9 +24,14 @@ function renderEventList(list) {
 
 
 async function fetchRefs() {
+    if (window.screen.availWidth >= 768 && window.screen.availWidth < 1280) {
+        eventsApiService.changeSize(21);
+    } else {
+        eventsApiService.changeSize(20);
+    }
 
     const list = await eventsApiService.fetchEvent({});
     const newFetchEventList = createsDownloadList(list);
     renderEventList(newFetchEventList);
-    getPage(eventsApiService, '', true);
+    //getPage(eventsApiService, '', true);
 }
